tests/submission/Infraware/Forms/contents/Forms/Not Use/input_min.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_min</h3>
	</p>
	
	<li>Min value of number is 1, you cannot submit this form.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='number' id='input_number'></p>
	</form>

	<script>
		
		var input_number = document.getElementById("input_number");
		input_number.min = "1";

		if (typeof(input_number.min) == "string") {
			if (input_number.min != "1") {
				test(function() {
					assert_true(false, "min attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("min attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
